+++ modules/node-groups/modules/node-group/variables.tf
@@ -141,7 +141,7 @@ variable "proximity_placement_group_id" {
 variable "max_pods" {
   description = "Maximum number of pods per node, this only apples to clusters using the AzureCNI."
   type        = number
-  nullable    = false
+  nullable    = true
 }
 
 variable "max_surge" {